fix(monitor/cli): false-positive supply alert on burns + fragile eligibility hash check
Three bug-hunt fixes:

* monitor: supply anomaly check used `!=`, so on_chain_supply < claims ×
  CLAIM_AMOUNT fired an "ALERT: Supply mismatch" on EVERY legitimate token
  burn (ZKMToken supports burn/burnFrom, which only ever reduce totalSupply).
  A sound system can never have supply ABOVE that line — that would be
  over-minting (a real exploit). Now alerts only on the impossible over-mint
  case and reports the gap as burned ZKM info-level, mirroring cmd_status.

* commands (cmd_fetch): the eligibility-list SHA-256 comparison was a raw
  `==` against the manifest hash, so a manifest field carrying a `0x` prefix,
  surrounding whitespace, or uppercase hex would falsely reject a VALID file
  with "SHA-256 mismatch". The KZG-SRS path already normalizes via
  `.trim().to_lowercase()`; this aligns the eligibility path to the same
  robust comparison (no-op for the expected clean lowercase hashes).

* circuits: drop a redundant `as u32` (clippy::unnecessary_cast) in
  test_measure_circuit_rows so `cargo clippy --workspace --all-targets --
  -D warnings` stays clean for first-party code.

+++ circuits/src/lib.rs
@@ -984,7 +984,7 @@ mod tests {
         let total = rows_used + blinding + 1; // rows that must fit in 2^k
 
         // Smallest k with 2^k >= total.
-        let min_k = (total as u64).next_power_of_two().trailing_zeros() as u32;
+        let min_k = (total as u64).next_power_of_two().trailing_zeros();
 
         let fits_k23 = total <= (1usize << 23);
         let fits_k24 = total <= (1usize << 24);
